Merged top modules
I start creating a new module that can replace:
Top
Latest
Last_updated
Rating.

Additional is also has a:
random function.
Different sort options ( like in the related module )

Todo is the template which is currently based on Top. It seems like they all are kind of the same but don`t know or i can merge it easily.